Refactor to com.davidmiguel package
