+++ lib/src/commonMain/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.davidmigloz.numberkeyboard"/>
+<manifest package="com.davidmiguel.numberkeyboard"/>
