--- lib/src/commonMain/kotlin/com/davidmigloz/numberkeyboard/data/NumberKeyboardFormat.kt
@@ -1,12 +1,0 @@
-package com.davidmigloz.numberkeyboard.data
-
-enum class NumberKeyboardFormat {
-    Normal,
-    Inverted,
-    Scrambled,
-    AlwaysScrambled;
-
-    companion object {
-        val Default = Normal
-    }
-}
