--- lib/src/commonMain/kotlin/com/davidmigloz/numberkeyboard/expect.kt
@@ -1,5 +1,0 @@
-package com.davidmigloz.numberkeyboard
-
-expect fun getDecimalSeparator(): Char
-expect fun getGroupingSeparator(): Char
-expect fun isCurrencySymbolInFront(): Boolean
