--- lib/src/commonMain/kotlin/com/davidmigloz/numberkeyboard/listener/NumberKeyboardClickedListener.kt
@@ -1,7 +1,0 @@
-package com.davidmigloz.numberkeyboard.listener
-
-interface NumberKeyboardClickedListener {
-    fun onNumberClicked(number: Int)
-    fun onLeftAuxButtonClicked()
-    fun onRightAuxButtonClicked()
-}
