--- lib/src/commonMain/kotlin/com/davidmigloz/numberkeyboard/listener/NumberKeyboardListener.kt
@@ -1,7 +1,0 @@
-package com.davidmigloz.numberkeyboard.listener
-
-import com.davidmigloz.numberkeyboard.data.NumberKeyboardData
-
-interface NumberKeyboardListener {
-    fun onUpdated(data: NumberKeyboardData)
-}
