+++ lib/src/commonMain/kotlin/com/davidmiguel/numberkeyboard/data/NumberKeyboardFormat.kt
@@ -1,0 +1,12 @@
+package com.davidmiguel.numberkeyboard.data
+
+enum class NumberKeyboardFormat {
+    Normal,
+    Inverted,
+    Scrambled,
+    AlwaysScrambled;
+
+    companion object {
+        val Default = Normal
+    }
+}
