+++ lib/src/commonMain/kotlin/com/davidmiguel/numberkeyboard/expect.kt
@@ -1,0 +1,5 @@
+package com.davidmiguel.numberkeyboard
+
+expect fun getDecimalSeparator(): Char
+expect fun getGroupingSeparator(): Char
+expect fun isCurrencySymbolInFront(): Boolean
