+++ lib/src/commonMain/kotlin/com/davidmiguel/numberkeyboard/listener/NumberKeyboardClickedListener.kt
@@ -1,0 +1,7 @@
+package com.davidmiguel.numberkeyboard.listener
+
+interface NumberKeyboardClickedListener {
+    fun onNumberClicked(number: Int)
+    fun onLeftAuxButtonClicked()
+    fun onRightAuxButtonClicked()
+}
