+++ lib/src/commonMain/kotlin/com/davidmiguel/numberkeyboard/listener/NumberKeyboardListener.kt
@@ -1,0 +1,7 @@
+package com.davidmiguel.numberkeyboard.listener
+
+import com.davidmiguel.numberkeyboard.data.NumberKeyboardData
+
+interface NumberKeyboardListener {
+    fun onUpdated(data: NumberKeyboardData)
+}
