--- sample/composeApp/src/androidMain/kotlin/com/davidmigloz/numberkeyboard/sample/MainActivity.kt
@@ -1,17 +1,0 @@
-package com.davidmigloz.numberkeyboard.sample
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-
-class MainActivity : ComponentActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        enableEdgeToEdge()
-        super.onCreate(savedInstanceState)
-        setContent {
-            App()
-        }
-    }
-}
