+++ sample/composeApp/src/androidMain/kotlin/com/davidmiguel/numberkeyboard/sample/MainActivity.kt
@@ -1,0 +1,17 @@
+package com.davidmiguel.numberkeyboard.sample
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+
+class MainActivity : ComponentActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        enableEdgeToEdge()
+        super.onCreate(savedInstanceState)
+        setContent {
+            App()
+        }
+    }
+}
