--- sample/composeApp/src/desktopMain/kotlin/com/davidmigloz/numberkeyboard/sample/main.kt
@@ -1,13 +1,0 @@
-package com.davidmigloz.numberkeyboard.sample
-
-import androidx.compose.ui.window.Window
-import androidx.compose.ui.window.application
-
-fun main() = application {
-    Window(
-        onCloseRequest = ::exitApplication,
-        title = "NumberKeyboard",
-    ) {
-        App()
-    }
-}
