+++ sample/composeApp/src/desktopMain/kotlin/com/davidmiguel/numberkeyboard/sample/main.kt
@@ -1,0 +1,13 @@
+package com.davidmiguel.numberkeyboard.sample
+
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.application
+
+fun main() = application {
+    Window(
+        onCloseRequest = ::exitApplication,
+        title = "NumberKeyboard",
+    ) {
+        App()
+    }
+}
