--- sample/composeApp/src/iosMain/kotlin/com/davidmigloz/numberkeyboard/sample/MainViewController.kt
@@ -1,5 +1,0 @@
-package com.davidmigloz.numberkeyboard.sample
-
-import androidx.compose.ui.window.ComposeUIViewController
-
-fun MainViewController() = ComposeUIViewController { App() }
