+++ sample/composeApp/src/iosMain/kotlin/com/davidmiguel/numberkeyboard/sample/MainViewController.kt
@@ -1,0 +1,5 @@
+package com.davidmiguel.numberkeyboard.sample
+
+import androidx.compose.ui.window.ComposeUIViewController
+
+fun MainViewController() = ComposeUIViewController { App() }
