+++ sample/iosApp/Configuration/Config.xcconfig
@@ -1,7 +1,7 @@
 TEAM_ID=
 
 PRODUCT_NAME=NumberKeyboard
-PRODUCT_BUNDLE_IDENTIFIER=com.davidmigloz.numberkeyboard.sample.NumberKeyboard$(TEAM_ID)
+PRODUCT_BUNDLE_IDENTIFIER=com.davidmiguel.numberkeyboard.sample.NumberKeyboard$(TEAM_ID)
 
 CURRENT_PROJECT_VERSION=1
 MARKETING_VERSION=1.0
